receptor-fastapi/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pyserial>=3.5
